# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TradingValueItem {
  /** 거래대금 순위 */
  rank: number;
  /** 종목코드 */
  ticker: string;
  /** 종목명 */
  name: string;
  /**
   * 업종 (없을 수 있음)
   * @nullable
   */
  sector?: string | null;
  /**
   * 시가총액
   * @nullable
   */
  marketCap?: number | null;
  /** 종가 */
  price: number;
  /** 전일 대비 */
  change: number;
  /** 등락률 (%) */
  changeRate: number;
  /** 거래량 (주) */
  volume: number;
  /** 거래대금 */
  tradingValue: number;
  /** 전체 시장 거래대금 대비 비중 (0~1) */
  tradingValueShare: number;
}

export interface TopTradingValueResponse {
  /** 조회 기준일 (YYYY-MM-DD) */
  date: string;
  /** 시장 코드 */
  market: string;
  /** 통화 코드 (KRW | USD) */
  currency: string;
  /** 시장 전체 거래대금 */
  marketTradingValue: number;
  /** 상위 N 종목 거래대금 합계 */
  topTradingValue: number;
  /** 상위 N 종목이 차지한 비중 (0~1) */
  topShare: number;
  items: TradingValueItem[];
}

export interface MarketSummary {
  date: string;
  market: string;
  currency: string;
  /** 시장 전체 거래대금 */
  marketTradingValue: number;
  /** 시장 전체 거래량 (주) */
  marketVolume: number;
  /** 거래 종목 수 */
  tickerCount: number;
  /** 상위 10종목 비중 (0~1) */
  top10Share: number;
  /** 1위 종목 비중 (0~1) */
  top1Share: number;
  /** 1위 종목코드 */
  topTicker: string;
  /** 1위 종목명 */
  topName: string;
  /** 1위 종목 거래대금 */
  topTradingValue: number;
  /** 상승 종목 수 (전일 대비 등락률 > 0) */
  advancers: number;
  /** 하락 종목 수 (전일 대비 등락률 < 0) */
  decliners: number;
  /** 보합 종목 수 */
  unchanged: number;
  /** 상승 종목 비율 (0~1) */
  advanceRate: number;
  /**
   * 서버 Yahoo 데이터 최종 갱신 시각 (ISO 8601)
   * @nullable
   */
  cacheRefreshedAt?: string | null;
}

export interface DailyTradingPoint {
  /** 영업일 (YYYY-MM-DD) */
  date: string;
  ticker: string;
  tradingValue: number;
}

export interface DailyTradingSeriesResponse {
  market: string;
  currency: string;
  from: string;
  to: string;
  items: DailyTradingPoint[];
}

export interface PeriodRankingItem {
  /** 기간 랭킹 순위 */
  rank: number;
  ticker: string;
  name: string;
  /** @nullable */
  sector?: string | null;
  /** 기간 점수: 일별로 (topN - dailyRank + 1) 의 합. topN 밖이면 0점 */
  score: number;
  /** 이론상 최대 점수 (모든 영업일에 1위 = topN × 영업일수) */
  maxScore: number;
  /** topN 안에 포함된 영업일 수 */
  daysInTop: number;
  /** topN 안일 때의 평균 일별 순위 */
  avgRank: number;
  /** 기간 중 도달한 최고 일별 순위 (작을수록 높음) */
  bestRank: number;
  /** 기간 동안 총 거래대금 */
  totalTradingValue: number;
}

export interface PeriodRankingResponse {
  market: string;
  currency: string;
  /** 조회 시작일 (YYYY-MM-DD, 영업일로 스냅됨) */
  from: string;
  /** 조회 종료일 (YYYY-MM-DD, 영업일로 스냅됨) */
  to: string;
  /** 기간 내 영업일 수 */
  daysInRange: number;
  /** 일별 점수 산정 기준 (상위 N) */
  topN: number;
  items: PeriodRankingItem[];
}

export type MarketParamParameter =
  (typeof MarketParamParameter)[keyof typeof MarketParamParameter];

export const MarketParamParameter = {
  kospi: "kospi",
  us: "us",
} as const;

export type GetTopTradingValueParams = {
  /**
   * 시장 코드 (kospi | us)
   */
  market?: MarketParamParameter;
  /**
   * 조회 기준일 (YYYY-MM-DD). 미지정 시 가장 최근 영업일 데이터 반환
   */
  date?: string;
  /**
   * 반환할 종목 개수 (기본값 10)
   */
  limit?: number;
};

export type GetMarketSummaryParams = {
  /**
   * 시장 코드 (kospi | us)
   */
  market?: MarketParamParameter;
  /**
   * 조회 기준일 (YYYY-MM-DD)
   */
  date?: string;
};

export type GetPeriodRankingParams = {
  /**
   * 시장 코드 (kospi | us)
   */
  market?: MarketParamParameter;
  /**
   * 기간 시작일 (YYYY-MM-DD). 미지정 시 가장 오래된 영업일
   */
  from?: string;
  /**
   * 기간 종료일 (YYYY-MM-DD). 미지정 시 가장 최근 영업일
   */
  to?: string;
  /**
   * 일별 점수 산정 기준 (기본 30, 즉 매일 상위 30위까지에 점수 부여)
   */
  topN?: number;
  /**
   * 반환할 종목 개수 (기본 10)
   */
  limit?: number;
};

export type GetDailyTradingSeriesParams = {
  /**
   * 시장 코드 (kospi | us)
   */
  market?: MarketParamParameter;
  /**
   * 시작일 (YYYY-MM-DD)
   */
  from: string;
  /**
   * 종료일 (YYYY-MM-DD)
   */
  to: string;
  /**
   * 쉼표로 구분된 종목코드 목록
   */
  tickers: string;
};

export type GetAvailableDatesParams = {
  /**
   * 시장 코드 (kospi | us)
   */
  market?: MarketParamParameter;
  /**
   * 반환할 일수 (기본 14)
   */
  count?: number;
};