# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 주식 거래대금 집중도 시각화 API (KOSPI, US)
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: markets
    description: 주식시장 거래대금 데이터
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /kospi/top-trading-value:
    get:
      operationId: getTopTradingValue
      tags: [markets]
      summary: 거래대금 상위 종목
      description: 특정 날짜의 종목을 거래대금 순으로 정렬하여 상위 N개 반환
      parameters:
        - $ref: "#/components/parameters/MarketParam"
        - name: date
          in: query
          required: false
          description: "조회 기준일 (YYYY-MM-DD). 미지정 시 가장 최근 영업일 데이터 반환"
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: "반환할 종목 개수 (기본값 10)"
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 거래대금 상위 종목 데이터
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopTradingValueResponse"
  /kospi/market-summary:
    get:
      operationId: getMarketSummary
      tags: [markets]
      summary: 시장 거래대금 요약
      description: 시장 전체 거래대금, 상위 종목 집중도, 거래 종목 수 등 요약 지표
      parameters:
        - $ref: "#/components/parameters/MarketParam"
        - name: date
          in: query
          required: false
          description: "조회 기준일 (YYYY-MM-DD)"
          schema:
            type: string
      responses:
        "200":
          description: 시장 요약
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketSummary"
  /kospi/period-ranking:
    get:
      operationId: getPeriodRanking
      tags: [markets]
      summary: 기간 거래대금 랭킹
      description: 지정한 기간 동안 일별 거래대금 상위 N에 포함된 빈도와 순위 가중치를 합산하여 산출한 종목 랭킹
      parameters:
        - $ref: "#/components/parameters/MarketParam"
        - name: from
          in: query
          required: false
          description: "기간 시작일 (YYYY-MM-DD). 미지정 시 가장 오래된 영업일"
          schema:
            type: string
        - name: to
          in: query
          required: false
          description: "기간 종료일 (YYYY-MM-DD). 미지정 시 가장 최근 영업일"
          schema:
            type: string
        - name: topN
          in: query
          required: false
          description: "일별 점수 산정 기준 (기본 30, 즉 매일 상위 30위까지에 점수 부여)"
          schema:
            type: integer
            default: 30
        - name: limit
          in: query
          required: false
          description: "반환할 종목 개수 (기본 10)"
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 기간 랭킹 데이터
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodRankingResponse"
  /kospi/daily-trading-series:
    get:
      operationId: getDailyTradingSeries
      tags: [markets]
      summary: 기간별 종목 일별 거래대금 시계열
      description: 지정한 기간 동안 특정 종목들의 일별 거래대금 시계열 반환
      parameters:
        - $ref: "#/components/parameters/MarketParam"
        - name: from
          in: query
          required: true
          description: "시작일 (YYYY-MM-DD)"
          schema:
            type: string
        - name: to
          in: query
          required: true
          description: "종료일 (YYYY-MM-DD)"
          schema:
            type: string
        - name: tickers
          in: query
          required: true
          description: "쉼표로 구분된 종목코드 목록"
          schema:
            type: string
      responses:
        "200":
          description: 일별 거래대금 시계열
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyTradingSeriesResponse"
  /kospi/available-dates:
    get:
      operationId: getAvailableDates
      tags: [markets]
      summary: 데이터 조회 가능 영업일
      description: 최근 영업일 목록 (시장별)
      parameters:
        - $ref: "#/components/parameters/MarketParam"
        - name: count
          in: query
          required: false
          description: "반환할 일수 (기본 14)"
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: 영업일 목록 (최신순)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
components:
  parameters:
    MarketParam:
      name: market
      in: query
      required: false
      description: "시장 코드 (kospi | us)"
      schema:
        type: string
        enum: [kospi, us]
        default: kospi
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    TradingValueItem:
      type: object
      properties:
        rank:
          type: integer
          description: "거래대금 순위"
        ticker:
          type: string
          description: "종목코드"
        name:
          type: string
          description: "종목명"
        sector:
          type: ["string", "null"]
          description: "업종 (없을 수 있음)"
        marketCap:
          type: ["number", "null"]
          description: "시가총액"
        price:
          type: number
          description: "종가"
        change:
          type: number
          description: "전일 대비"
        changeRate:
          type: number
          description: "등락률 (%)"
        volume:
          type: number
          description: "거래량 (주)"
        tradingValue:
          type: number
          description: "거래대금"
        tradingValueShare:
          type: number
          description: "전체 시장 거래대금 대비 비중 (0~1)"
      required:
        - rank
        - ticker
        - name
        - price
        - change
        - changeRate
        - volume
        - tradingValue
        - tradingValueShare
    TopTradingValueResponse:
      type: object
      properties:
        date:
          type: string
          description: "조회 기준일 (YYYY-MM-DD)"
        market:
          type: string
          description: "시장 코드"
        currency:
          type: string
          description: "통화 코드 (KRW | USD)"
        marketTradingValue:
          type: number
          description: "시장 전체 거래대금"
        topTradingValue:
          type: number
          description: "상위 N 종목 거래대금 합계"
        topShare:
          type: number
          description: "상위 N 종목이 차지한 비중 (0~1)"
        items:
          type: array
          items:
            $ref: "#/components/schemas/TradingValueItem"
      required:
        - date
        - market
        - currency
        - marketTradingValue
        - topTradingValue
        - topShare
        - items
    MarketSummary:
      type: object
      properties:
        date:
          type: string
        market:
          type: string
        currency:
          type: string
        marketTradingValue:
          type: number
          description: "시장 전체 거래대금"
        marketVolume:
          type: number
          description: "시장 전체 거래량 (주)"
        tickerCount:
          type: integer
          description: "거래 종목 수"
        top10Share:
          type: number
          description: "상위 10종목 비중 (0~1)"
        top1Share:
          type: number
          description: "1위 종목 비중 (0~1)"
        topTicker:
          type: string
          description: "1위 종목코드"
        topName:
          type: string
          description: "1위 종목명"
        topTradingValue:
          type: number
          description: "1위 종목 거래대금"
        advancers:
          type: integer
          description: "상승 종목 수 (전일 대비 등락률 > 0)"
        decliners:
          type: integer
          description: "하락 종목 수 (전일 대비 등락률 < 0)"
        unchanged:
          type: integer
          description: "보합 종목 수"
        advanceRate:
          type: number
          description: "상승 종목 비율 (0~1)"
        cacheRefreshedAt:
          type: ["string", "null"]
          description: "서버 Yahoo 데이터 최종 갱신 시각 (ISO 8601)"
      required:
        - date
        - market
        - currency
        - marketTradingValue
        - marketVolume
        - tickerCount
        - top10Share
        - top1Share
        - topTicker
        - topName
        - topTradingValue
        - advancers
        - decliners
        - unchanged
        - advanceRate
    DailyTradingPoint:
      type: object
      properties:
        date:
          type: string
          description: "영업일 (YYYY-MM-DD)"
        ticker:
          type: string
        tradingValue:
          type: number
      required:
        - date
        - ticker
        - tradingValue
    DailyTradingSeriesResponse:
      type: object
      properties:
        market:
          type: string
        currency:
          type: string
        from:
          type: string
        to:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DailyTradingPoint"
      required:
        - market
        - currency
        - from
        - to
        - items
    PeriodRankingItem:
      type: object
      properties:
        rank:
          type: integer
          description: "기간 랭킹 순위"
        ticker:
          type: string
        name:
          type: string
        sector:
          type: ["string", "null"]
        score:
          type: number
          description: "기간 점수: 일별로 (topN - dailyRank + 1) 의 합. topN 밖이면 0점"
        maxScore:
          type: number
          description: "이론상 최대 점수 (모든 영업일에 1위 = topN × 영업일수)"
        daysInTop:
          type: integer
          description: "topN 안에 포함된 영업일 수"
        avgRank:
          type: number
          description: "topN 안일 때의 평균 일별 순위"
        bestRank:
          type: integer
          description: "기간 중 도달한 최고 일별 순위 (작을수록 높음)"
        totalTradingValue:
          type: number
          description: "기간 동안 총 거래대금"
      required:
        - rank
        - ticker
        - name
        - score
        - maxScore
        - daysInTop
        - avgRank
        - bestRank
        - totalTradingValue
    PeriodRankingResponse:
      type: object
      properties:
        market:
          type: string
        currency:
          type: string
        from:
          type: string
          description: "조회 시작일 (YYYY-MM-DD, 영업일로 스냅됨)"
        to:
          type: string
          description: "조회 종료일 (YYYY-MM-DD, 영업일로 스냅됨)"
        daysInRange:
          type: integer
          description: "기간 내 영업일 수"
        topN:
          type: integer
          description: "일별 점수 산정 기준 (상위 N)"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PeriodRankingItem"
      required:
        - market
        - currency
        - from
        - to
        - daysInRange
        - topN
        - items