# lib/api-zod/src/generated/types/dailyTradingPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */

export interface DailyTradingPoint {
  /** 영업일 (YYYY-MM-DD) */
  date: string;
  ticker: string;
  tradingValue: number;
}