# lib/api-zod/src/generated/types/dailyTradingSeriesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */
import type { DailyTradingPoint } from "./dailyTradingPoint";

export interface DailyTradingSeriesResponse {
  market: string;
  currency: string;
  from: string;
  to: string;
  items: DailyTradingPoint[];
}