# lib/api-zod/src/generated/types/periodRankingItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */

export interface PeriodRankingItem {
  /** 기간 랭킹 순위 */
  rank: number;
  ticker: string;
  name: string;
  /** @nullable */
  sector?: string | null;
  /** 기간 점수: 일별로 (topN - dailyRank + 1) 의 합. topN 밖이면 0점 */
  score: number;
  /** 이론상 최대 점수 (모든 영업일에 1위 = topN × 영업일수) */
  maxScore: number;
  /** topN 안에 포함된 영업일 수 */
  daysInTop: number;
  /** topN 안일 때의 평균 일별 순위 */
  avgRank: number;
  /** 기간 중 도달한 최고 일별 순위 (작을수록 높음) */
  bestRank: number;
  /** 기간 동안 총 거래대금 */
  totalTradingValue: number;
}