# lib/api-zod/src/generated/types/periodRankingResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */
import type { PeriodRankingItem } from "./periodRankingItem";

export interface PeriodRankingResponse {
  market: string;
  currency: string;
  /** 조회 시작일 (YYYY-MM-DD, 영업일로 스냅됨) */
  from: string;
  /** 조회 종료일 (YYYY-MM-DD, 영업일로 스냅됨) */
  to: string;
  /** 기간 내 영업일 수 */
  daysInRange: number;
  /** 일별 점수 산정 기준 (상위 N) */
  topN: number;
  items: PeriodRankingItem[];
}