# lib/api-zod/src/generated/types/topTradingValueResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */
import type { TradingValueItem } from "./tradingValueItem";

export interface TopTradingValueResponse {
  /** 조회 기준일 (YYYY-MM-DD) */
  date: string;
  /** 시장 코드 */
  market: string;
  /** 통화 코드 (KRW | USD) */
  currency: string;
  /** 시장 전체 거래대금 */
  marketTradingValue: number;
  /** 상위 N 종목 거래대금 합계 */
  topTradingValue: number;
  /** 상위 N 종목이 차지한 비중 (0~1) */
  topShare: number;
  items: TradingValueItem[];
}