# lib/api-zod/src/generated/types/tradingValueItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 주식 거래대금 집중도 시각화 API (KOSPI, US)
 * OpenAPI spec version: 0.1.0
 */

export interface TradingValueItem {
  /** 거래대금 순위 */
  rank: number;
  /** 종목코드 */
  ticker: string;
  /** 종목명 */
  name: string;
  /**
   * 업종 (없을 수 있음)
   * @nullable
   */
  sector?: string | null;
  /**
   * 시가총액
   * @nullable
   */
  marketCap?: number | null;
  /** 종가 */
  price: number;
  /** 전일 대비 */
  change: number;
  /** 등락률 (%) */
  changeRate: number;
  /** 거래량 (주) */
  volume: number;
  /** 거래대금 */
  tradingValue: number;
  /** 전체 시장 거래대금 대비 비중 (0~1) */
  tradingValueShare: number;
}